Add auth audit attempt events to the SAML IdP. (#935)
The SAML IdP now emits auth attempt events when succeeding or failing to use
the IdP. We emit the events in a few places because there are a few places
that an auth attempt can terminate. These are controlled by the crewjam
SAML library, which is out of our control.

* At the end of a GetSession attempt, which can have a success or failure
  value.
* When failing to find a service provider, which terminates an auth attempt.
  This is only emits a failure.

I've additionally cleaned up a few of the tests which had common patterns.

Source-commit: gravitational/teleport.e@1e91341d86a3f9c58c1f6300140afe4f9c452a42
Source-pull-requests: gravitational/teleport.e#935