Sign SAML responses on the auth server. (#957)
SAML responses are now signed on the auth server. By doing this, we prevent
the SAML IdP located on the proxy server from signing the responses itself,
and will eliminate the need for the SAML IdP to have any private key material.
This introduces a new SAML IdP signing service which lives on the auth server
that will handle signing.

Source-commit: gravitational/teleport.e@725bee6a025e6bf9429ec0ee15be5dfafd227de2
Source-pull-requests: gravitational/teleport.e#957