Create the auth server in the startup function.
In enterprise, the auth plugins require access to the backend. Currently,
when the TLS server is created, auth plugins are registered prior to the
`TeleportProcess` being returned from `service.NewTeleport`. This results in
the auth plugins needing to lazily evaluate the getter for the `backend`,
which subsequently makes an awakward interface for enterprise plugins that
rely on backend storage.

This creates the TLS server in the first critical function for the auth
server, which will be executed by `service.Run` after the `TeleportProcess`
has been returned. This ensures that the backend is available when the
enterprise auth plugins are registered, which will simplify enterprise
auth plugin implementations.