docs: mention how to get the correct API version (#22811)
Our API module is not currently versioned to align with Teleport
major versions, but it's important for developers building integrations
to use a verison of the API that is compatible with their cluster.

Include a short snippet that can be used to identify the correct
version string to place in go.mod given a particular tag.