Change kube_servers to in memory cache feeded by a watcher
This PR introduces a in memory cache for `kube_servers` that watches the
creation/deletion of those objects and keeps them organized into memory to speed
up the proxy process and to avoid unmarshaling servers per request.

The Kubernetes Service uses the registered clusters and does not need
the `kube_server` watcher to operate.

Fixes #22598