Provide more context in the docs intro page (#22517)
* Provide more context in the docs intro page

Closes #22469

The current docs landing page does not include enough context for new
readers. After the general description of Teleport, the rest of the
landing page is series of links, with little guidance for what a new
user should do first (or what an intermediate user should do after
completing the initial setup steps).

This change takes a more narrative approach to the landing page,
briefly describing the content of each top-level docs section and
putting it in context with the rest of the docs. This also helps users
find the right top-level docs section for their needs, since otherwise
the only way to do this is to visit each section.

This change is also more opinionated about which guide new users should
begin with. It recommends the guide to getting started with Teleport
Cloud, since this doesn't require setting up a Teleport cluster and
helps a user register a resource quickly.

The landing page links to the introduction page of each top-level docs
section. If a section does not have an introduction page, this change
adds one.

* Respond to zmb3 feedback

Reorder list items

* Linter fixes

* Respond to alexfornuto PR feedback