Add dummy MacOS build workflow (#989)
Add a dummy workflow that can be merged to master so that a development
branch can iterate on the actual real workflow until approved for
merging. GitHub Actions will not run a `workflow_dispatch` workflow
unless it exists in some form on the default branch. The version on the
development branch can be completely different, but without the named
workflow on the default branch, you are unable to trigger it at all.

Source-commit: gravitational/teleport.e@c0d7022d4527eb5df2fb4662a5eb18ebe78b8589
Source-pull-requests: gravitational/teleport.e#989