Handle getBackend() or backend argument for plugins.
Now that the enterprise code no longer needs this lazily evaluated backend
function, the plugins local services has been temporarily changed to handle
both `getBackend()` and a regular `backend.Backend` argument until the
reference to it can be removed in enterprise. Then, a follow on will be
submitted to only support the `backend.Backend` argument.