Update `crypto` to `0.6.0` and remove our replacement in `api` (#23131)
* Update `crypto` in api

It was missed that cyrpto is being replaced in our `api` as well.
This commit updates the version to match the one in the repo root.

* Remove replacement for crypto in `api`

Replacement directives only function when defined at the main project level, they are not inherited.  For that reason I believe here we need to only depend on the updated crypto package, rather than update our replacement.

---------

Co-authored-by: Jakub Nyckowski <jakub.nyckowski@goteleport.com>