Add Dummy GHA workflow for rebuild OCI images (#999)
Add a dummy workflow that can be merged to master so that a development
branch can iterate on the actual real workflow until approved for
merging. GitHub Actions will not run a `workflow_dispatch` workflow
unless it exists in some form on the default branch.

The version on the development branch can be completely different, 
but without the named workflow on the default branch, you are unable
to trigger it at all.

Source-commit: gravitational/teleport.e@b8a6036fa77ad19d54ac720e611ef9c1c7d8b89f
Source-pull-requests: gravitational/teleport.e#999