Prepare enterprise for enterprise only services. (#1001)
The upcoming Okta service is going to be the first service that can run
independently of the auth service in enterprise. This does some refactoring
to the enterprise process startup so that it should be easier to add in the
Okta service (and other enterprise services) that run independently of the
auth service.

Source-commit: gravitational/teleport.e@60e6efd86bbe79442a83a64e32aef99441d58e95
Source-pull-requests: gravitational/teleport.e#1001