[v12] build.assets Dockerfiles: Remove unnecessary ENV NODE_URL, pass fsSL to curl (#23188)
* build.assets Dockerfiles: Remove unnecessary ENV NODE_URL

NODE_URL is being redefined within the RUN instruction anyway. We suspect
it might be causing problems because sometimes the logs from build failures
suggest that the NODE_URL export was either ignored or ${NODE_URL} passed
to curl reads ENV NODE_URL and not the env var set within the shell.

* Pass fsSl flags to curl