Fix docs metioning upserting connectors without secrets
When upserting Github/SAML/OIDC connectors we must upsert them with
their original secrets. The exporter command didn't exported the
connectors with their original secrets and the upsert replaced them.

The result is that no one is able to login to Teleport using those
connectors because of the missing secrets.

This PR corrects the instructions used to extract the connectors and
their secrets.

It also fixes a documentation reference to incorrect resources.