[v11] Fix incorrect report of active sessions (#23445)
* Fix incorrect report of active sessions

The session tracker didn't validate the current state of the tracked session and allowed out-of-order events to replace the session state if the session was already marked as terminated.

This resulted in the session being marked as active but the expiry TTL was equal to the same as a terminated session which eventually resulted in the session being active for days due to the improper deletion propagation that DynamoDB has.

This PR checks the state and if the session is already marked as closed, it denies any update.

* add test case