Add `srv.ConnectionMonitor` to unify connection monitoring logic (#23465)
Wraps the common use of `srv.StartMonitor` into a dedicated type
so that the logic doesn't need to exist once per protocol. Both
apps and db have been migrated to use the
`srv.ConnectionMonitor` - kube and ssh could also be refactored in
the future. Windows desktops have a slightly different usage but
could also be converted with a slight tweak to the new connection
monitor.