[v11] Fetch kubernetes git version with disabled service account (#23558)
* Fetch kubernetes git version with disabled service account

This commit ensures that the git version of the kubernetes cluster where
agent is (potentially) running on can be retrieved even if the service
account is disabled at the pod.

Our helm charts have `automountServiceAccountToken: true`, and this
allows the current implementation (that uses `InClusterConfig`) to work.
However, this is not the case e.g. in auth and proxy pods running on
Teleport cloud which have `automountServiceAccountToken: false`.

* Make var const

Co-authored-by: Alan Parra <alan.parra@goteleport.com>

* Remove `log.Debugf` calls

* Make var const

* Check if git version is empty

* Do not create http request if KUBERNETES_SERVICE_{HOST,PORT} env vars are empty

* Remove unused import

---------

Co-authored-by: Alan Parra <alan.parra@goteleport.com>