[v11] Increase DialTimeout when testing SSH Connection Diagnostics (#23634)
* Increase DialTimeout when testing SSH Connection Diagnostics

Depending on the system, 1 second might not be enough to complete the test.
Specifically for darwin/amd64 on either Intel or M1 hosts

I was able to get a failure by using amd64 on an M1 host.
The error is: context canceled
Which indicates a context timeout.

After setting this value to 5, I was able to get a PASS on the test.
The test time didn't increase significantly.
So, it seems a nice trade off.

* add caveat about reducing the DialTimeout to lower values