Allow user impersonation from Teleport proxies (#23311)
This PR introduces user impersonation from connections created by Teleport proxies. The goal is to allow proxies to forward the user identity that initiated the connection without signing a new certificate with the user's identity.

HTTP requests originated from Proxies with `Teleport-Impersonate-User` header allows the proxy to impersonate the identity carried in the value.

Certificates without `SytemRole=Proxy` are denied to impersonate users and the connection is automatically terminated.

For IP Pinning, the client IP address is transferred using the `Teleport-Real-User-IP` header.

Since the connection originated from a valid certificate belonging to a Teleport proxy, the identity and IP address headers don't need to be signed because the Teleport upstream service trusts the proxy certificate.

Part of #22533