Fix missing `kube_service` when old clients dial to T13
When a Teleport 12 client dials to Auth it expects the
`kube_service` resource type to exist in Teleport Auth Server.
Since #22663, Teleport no longer supports `kube_service` types but due
to a bad rebase this commit was lost and without it forces that any
Teleport 12 client when dialing to Teleport 13 auth fails because it
tries to execute the fallback method of `GetKubernetesServers` which calls `GetKubernetesServices`.