Return exit code from SFTP subsystem (#23626)
Go SFTP doesn't return error code as mentioned here https://github.com/pkg/sftp/issues/321. As a consequence of that `scp` returns error code 1 if no code wa returned for the server, which makes it look like something went wrong even on successful transfer.

This PR adds the return code to the SFTP subsystem, which fixes the described behavior.

Closes https://github.com/gravitational/teleport/issues/20863

Note: The error code is not returned when using OpenSSH server and client. This PR only fixes the behavior when Teleport is used as a server. Here is an issue tro trace it https://github.com/gravitational/teleport/issues/23625