[v11] Add SFTP subsystem fails note to server access FAQ (#23361)
* Add SFTP subsystem fails note to server access FAQ

In many cases we see people trying to use `tsh` v11 with Teleport Node v9. In this combination `ssh: subsystem request failed` which is very confusing for users. As we're not able to make sure that all Teleport installations are up-to date, we at least can we mention the most common solution in our FAQ.

* Update docs/pages/faq.mdx

Co-authored-by: Alex Fornuto <alex.fornuto@goteleport.com>

---------

Co-authored-by: Alex Fornuto <alex.fornuto@goteleport.com>