test: run login rule controller tests with an enterprise auth client (#1027)
This commit runs the tests for the Login Rule controller of the Teleport
Kubernetes operator from a test file in the enterprise repo. The tests
must be run from the enterprise repo so that a client to an enterprise
auth server can be passed in. Login rules are not supported in an OSS
auth server.

Source-commit: gravitational/teleport.e@4cc459face99d0970f2edcf624fc5f7ddc35a58e
Source-pull-requests: gravitational/teleport.e#1027