release: Add a workflow for building MacOS binaries (#1019)
Add a callable workflow and another dispatchable workflow to build the
MacOS binaries, sign and notarize them and then upload to the release
server.

This follows the same split as used in the linux ARM64 builds - a build
workflow that is called to do the actual building, and a release
workflow that can be dispatched either via the GitHub UI or via the
command line, to be integrated into Drone. Later the release workflow
will be triggered on a tag, and the build workflow triggered on a push.

This depends on PR gravitational/teleport#23407 to be merged first as it
lays the groundwork for this workflow to work (setting up signing keys
and propagating to the build scripts).

This currently contains a bunch of yaml copied and pasted from the
Linux build/release workflows. Once the full Mac build/release workflow
has landed, we'll see what the opportunities are for extracting commonality
into their own callable workflows. There are presently some clearly
common steps, but it also looks like there may be some small differences
too, so when we see the final shape of this, we can parameterize
accordingly.

Source-commit: gravitational/teleport.e@a2241662f4c9615d9b8ee536ed520561767c991f
Source-pull-requests: gravitational/teleport.e#1019