Clarify wording of Connect's Telemetry FAQ (#23413) (#23739)
This clarifies the wording of Connect's Telemetry FAQ which implied
that it's impossible in all cases to deanonymize event fields, which
may not be true if the cluster's internal datastore is compromised.

This attempts to paint a more accurate picture of the deanonymization
threat model, since ultimately the cluster's random ID can be used in
a dictionary-style attack if known or leaked.