[v10] Reduce time spent setting ssh session envs  (#23832)
* Reduce time spent setting ssh session envs

`tsh` sets a number of environment variables when setting up the
users session. Each key value pair is transmitted one at a time
in a "env" ssh request, which adds a num envs * RTT of additional
latency per session.

This introduces a new `envs@goteleport.com` request which sets
multiple environment variables in a single ssh request, which
reduces the amount of time spent setting envs down to the RTT of
a single ssh request. In order to ensure backward compat and
interoperability with OpenSSH, if the server does not recognize
the `envs@goteleport.com` request the ssh client will resort to
sending individual "env" requests.

* address feedback

* fix: use a single timer for fallback requests in tests

Co-authored-by: Alan Parra <alan.parra@goteleport.com>

* fix: remove extra whitespace

Co-authored-by: Alan Parra <alan.parra@goteleport.com>

* fix: gci

---------

Co-authored-by: Alan Parra <alan.parra@goteleport.com>