Backport #23405 to branch/v12 (#23883)
* Edit cluster joining info in Access Request docs

See #21305

Edit Access Request plugin guides to remove the options to connect to
the Auth Service directly. This simplifies the guides and helps us
standardize the docs around connecting services via the Proxy Service.

Also make the Jira guide more consistent with other Access Request
plugin guides (this is not an attempt to refresh the guide, which is
still out of date, but will help us refresh the guide when the time
comes).

* Fix linter issues

* Respond to PR feedback