Add App/AppServer SubKind.
Apps and AppServers now have SubKinds to denote different types of app
servers that can exist. With the addition of the Okta service, we now have
AppServers that may not be acted on by Application services. Therefore, we
should use SubKind to differentiate the different types of app servers that
may exist.