Backport #22817 to branch/v12 (#23881)
* Add a guide to using the API for auto-discovery

See #19716

Adding resources is a popular use of Teleport's API, so I have added a
guide to using the API for syncing resources to a service discovery API
using an example of a local Docker setup.

* Respond to zmb3 feedback

* Use the examples directory for the code

This way, users can have a compilable example before they start working
through the guide.

* Respond to alexfornuto feedback

* Respond to PR feedback

* Run make fix-license