release: Build MacOS packages in GHA workflow (#1050)
* build: Build Teleport Connect for MacOS

Build Teleport Connect for MacOS, but only for a push build, not a
release build. The release build requires the tsh.app package which we
do not yet build. This new build step is used by the push build to
ensure that we can build Teleport Connect.

* release: Build MacOS Teleport package

Add a step to the MacOS build pipeline to build the Teleport package.
Copy the OSS and Enterprise packages to the artifacts directory so they
are later uploaded to the release server.

Ensure the signing key IDs are exported before running
`build-package.sh`. Previously these were exported by the top-level
Makefile, however that changed when we changed the enterprise Makefile
to include the `darwin-signing.mk` file instead. This export should have
been added at that time.

Pass the bundle ID to the `build-package.sh` script so we use the dev
package name when creating the Teleport Enterprise package.

* release: Build MacOS tsh package

Add a step to the MacOS build pipeline to build the tsh package. Copy
the OSS and Enterprise packages to the artifacts directory so they are
later uploaded to the release server.

* release: Fix registration of release artifacts for all products

Fix the registration of the release products, appropriately setting the
description and products. The initial version was just for Teleport
only, and not the packages.

tsh and Teleport Connect need to have the one binary registered for OSS
and Enterprise as there are not separate binaries for each edition.

The tarball, teleport package, tsh package and Teleport Connect dmg all
have different descriptions.

Source-commit: gravitational/teleport.e@4cd1a5f89077850d95f2001128e6d32adcac2bed
Source-pull-requests: gravitational/teleport.e#1050