Enforce Kubernetes license when creating Kube resources (#23755)
This PR prevents the registration of Kubernetes clusters in Teleport
Auth server when the auth isn't licensed for Kubernetes access.

This is required because with the new credentials forwarding mechanism,
services won't call Auth server `ProcessKubeCSR` endpoint and thus we
cannot enforce the license.

This change prevents any agent to register its `KubeServers` in Auth
server if the server isn't properly licensed to Kube. It only affects
enterprise users because the open source version of Auth is licensed
for Kubernetes access.

Part of #22533