Remove the Access Controls FAQ
Fixes #11829

Teleport's Access Controls section is pretty sprawling, with 8 top-level
subsections and 46 pages. The Access Controls FAQ, meanwhile, is
unusually short, with only three questions. This can become misleading,
since there's a good chance that any information someone is looking for
in the FAQ will not exist there.

This change removes the Access Controls FAQ. If an answer found in the
FAQ does not exist explicitly elsewhere in the docs, I have added it.