Fix panic for when `/web/launch` is requested
When this path is requested a panic occurs due to the assumption 3 lines down of how many parts are done after the `Split`.  Without a trailing `/` there will only be 2 parts leading to an array bounds panic.

This does not appear to be a significantly impacting issue, but this one line fix seems trivial and safe to me.